+++ src/test/java/io/fabric8/maven/docker/config/BuildImageConfigurationTest.java
@@ -17,6 +17,7 @@ package io.fabric8.maven.docker.config;
  */
 
 import java.io.File;
+import java.io.IOException;
 
 import io.fabric8.maven.docker.util.Logger;
 import mockit.Expectations;
@@ -55,6 +56,7 @@ public class BuildImageConfigurationTest {
         config.initAndValidate(logger);
         assertTrue(config.isDockerFileMode());
         assertEquals(config.getDockerFile(),new File("src/docker/Dockerfile"));
+        assertEquals(config.getContextDir(),new File("src/docker"));
     }
 
     @Test
@@ -65,6 +67,7 @@ public class BuildImageConfigurationTest {
         config.initAndValidate(logger);
         assertTrue(config.isDockerFileMode());
         assertEquals(config.getDockerFile(),new File("src/docker/Dockerfile"));
+        assertEquals(config.getContextDir(),new File("src/docker"));
     }
 
     @Test
@@ -85,6 +88,53 @@ public class BuildImageConfigurationTest {
                 dockerFileDir("/tmp/").
                 dockerFile("/Dockerfile").build();
         config.initAndValidate(logger);
+    }
+
+    @Test
+    public void contextDir() {
+        BuildImageConfiguration config =
+                new BuildImageConfiguration.Builder().
+                        contextDir("target").build();
+        config.initAndValidate(logger);
+        assertEquals(new File("target"), config.getContextDir());
+    }
+
+    @Test
+    public void contextDirAndDockerfile() {
+        BuildImageConfiguration config =
+                new BuildImageConfiguration.Builder().
+                        dockerFile("src/docker/Dockerfile").
+                        contextDir("target").build();
+        config.initAndValidate(logger);
+        assertEquals(new File("target/src/docker/Dockerfile"), config.getDockerFile());
+        assertEquals(new File("target"), config.getContextDir());
+    }
+
+    @Test
+    public void contextDirAndDockerfileDir() {
+        BuildImageConfiguration config =
+                new BuildImageConfiguration.Builder().
+                        dockerFileDir("src/docker").
+                        contextDir("target").build();
+        config.initAndValidate(logger);
+        assertEquals(new File("target/Dockerfile"), config.getDockerFile());
+        assertEquals(new File("target"), config.getContextDir());
+    }
+
+    @Test
+    public void contextDirAndAbsoluteDockerfile() throws IOException {
+        File tempDockerFile = File.createTempFile("Dockerfile", "");
+        tempDockerFile.deleteOnExit();
+        BuildImageConfiguration config = new BuildImageConfiguration.Builder()
+                .dockerFile(tempDockerFile.getAbsolutePath())
+                .contextDir("target")
+                .build();
+
+        // If contextDir is given and the dockerFile is an absolute path.
+        // The Dockerfile should then be copied over.
+        config.initAndValidate(logger);
+        assertEquals(new File(tempDockerFile.getAbsolutePath()), config.getDockerFile());
+        assertEquals(new File("target"), config.getContextDir());
     }
 
     @Test